% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expressionData}
\alias{expressionData}
\title{Expression values of genes in tissues.}
\format{An object of class \code{data.frame} with 19613 rows and 35 columns.}
\usage{
expressionData
}
\description{
A dataset containing normalized gene expression values in different tissues.
}
\keyword{datasets}
