% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expressionValue}
\alias{expressionValue}
\title{Normalization Method used to process data}
\format{A data frame with 53940 rows and 10 variables:}
\usage{
expressionValue
}
\description{
Normalization Method used to process data
}
\keyword{datasets}
