% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueEnrich.R
\name{teGeneRetrieval}
\alias{teGeneRetrieval}
\title{Calculate tissue-specific genes by using the algorithm from the Human Protein Atlas project}
\usage{
teGeneRetrieval(expressionData, foldChangeThreshold = 5,
  maxNumberOfTissues = 7, expressedGeneThreshold = 1)
}
\arguments{
\item{expressionData}{A dataframe object containing gene expression values (Rows are genes and Tissues are columns).}

\item{foldChangeThreshold}{A numeric Threshold of fold change, default 5.}

\item{maxNumberOfTissues}{A numeric Maximum number of tissues in a group for group enriched genes, default 7.}

\item{expressedGeneThreshold}{A numeric Minimum gene expression cutoff for the gene to be called as expressed, default 1.}
}
\value{
A data frame object with three columns, Gene, Tissues, and Enrichment group of the gene in the given tissue.
}
\description{
This teGeneRetrieval function divides the genes into tissue-specific groups, based on the algorithm
used in Human Protein Atlas project. It takes a gene expression matrix as input (rows as genes and columns as tissue)
and classifies the genes into tissue-specific groups. More detail about the gene groups is provided in the vignette.
}
\examples{
data<-system.file("extdata", "test.expressiondata.txt", package = "TissueEnrich")
expressionData<-read.table(data,header=TRUE,row.names=1,sep='\\t')
TSgenes<-teGeneRetrieval(expressionData)
head(TSgenes)
}
\author{
Ashish Jain, Geetu Tuteja
}
