% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueEnrich.R
\name{teGenesRetrieval}
\alias{teGenesRetrieval}
\title{Calculation of tissue specific genes by using the algorithm from human protein atlas project.}
\usage{
teGenesRetrieval(expressionData, foldChangeThreshold = 5,
  maxNumberOfTissues = 7, expressedGeneThreshold = 1)
}
\arguments{
\item{expressionData}{A dataframe object containing gene expression values (Rows are genes and Tissues are columns) .}

\item{foldChangeThreshold}{A number. Threshold of fold change, default 5.}

\item{maxNumberOfTissues}{A number. Maximum number of tissues in a group for group enriched genes, default 7.}

\item{expressedGeneThreshold}{A number. Minimum gene expression cutoff for the gene to be called as expressed, default 1.}
}
\value{
A data frame object with three columns, Gene, Tissues, and the enrichment group of the gene in the given tissue.
}
\description{
Calculation of tissue specific genes by using the algorithm from human protein atlas project.
}
\examples{
data<-system.file("extdata", "combined-proteincodingGenedataCombine.txt", package = "TissueEnrich")
#expressionData<-read.table(data,header=TRUE,row.names=1,sep='\\t')
#TSgenes<-teGenesRetrieval(expressionData)
#head(TSgenes)
}
\author{
Ashish Jain, Geetu Tuteja
}
