% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissueDetails}
\alias{tissueDetails}
\title{Tissue Details}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{RName}{tissue name in R, character}
  \item{Sample}{Number of sample in study, integer}
  \item{TissueName}{tissue name, character}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
tissueDetails
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
