% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueEnrich.R
\name{tissueSpecificGeneEnrichment}
\alias{tissueSpecificGeneEnrichment}
\title{Calculation of tissue specific gene enrichment using hypergeometric test}
\usage{
tissueSpecificGeneEnrichment(inputGenes, tissueSpecificGenes = NULL,
  dataset = c("Human Protein Atlas", "GTEx combine", "GTEx sub-tissue",
  "Mouse ENCODE"), organism = c("Homo Sapiens", "Mus Musculus"),
  tissueSpecificGeneType = c("All", "Tissue-Enriched", "Tissue-Enhanced",
  "Group-Enriched"), geneSymbol = c("EnsemblId", "Gene Symbol"),
  isHomolog = FALSE)
}
\arguments{
\item{inputGenes}{A vector containing the input genes.}

\item{tissueSpecificGenes}{A dataframe object. Output from `tissueSpecificGenesRetrieval` function, default NULL.}

\item{dataset}{A String describing the dataset to be used for enrichment, Default "Human Protein Atlas". Other sources "GTEx combine","GTEx sub-tissue","Mouse ENCODE". Not required when providing \code{tissueSpecificGenes}.}

\item{organism}{A String describing the organism, default "Homo Sapiens", other "Mus Musculus". Not required when providing \code{tissueSpecificGenes}.}

\item{tissueSpecificGeneType}{A String describing type of tissue specific genes to be used, default "All", others "Tissue-Enriched","Tissue-Enhanced","Group-Enriched".}

\item{geneSymbol}{Type of gene symbol, default "Gene Symbol", other "EnsemblId".}

\item{isHomolog}{Flag for usage of orthologus genes, default FALSE. Not required when providing \code{tissueSpecificGenes}.}
}
\value{
A data frame object with three columns, Gene, Tissues, and the enrichment group of the gene in the given tissue.
}
\description{
Calculation of tissue specific gene enrichment using hypergeometric test
}
\examples{
library(tidyverse)
data<-read.table("humanProteinAtlasV18.tsv",header = T,sep = "\\t")
spreadData<-data \%>\% spread(key = Sample,value= Value)
row.names(spreadData)<-spreadData[,1]
humanProteinAtlas<-spreadData[gene,c(4:ncol(spreadData))]
TSgenes<-tissueSpecificGenesRetrieval(humanProteinAtlas)
head(TSgenes)
}
