% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueEnrich.R
\name{tissueSpecificGeneEnrichmentCustom}
\alias{tissueSpecificGeneEnrichmentCustom}
\title{Calculation of tissue specific gene enrichment using hypergeometric test for custom datasets}
\usage{
tissueSpecificGeneEnrichmentCustom(inputGenes = NULL,
  tissueSpecificGenes = NULL, tissueSpecificGeneType = 1,
  multiHypoCorrection = TRUE)
}
\arguments{
\item{inputGenes}{A vector containing the input genes.}

\item{tissueSpecificGenes}{A dataframe object. Output from `tissueSpecificGenesRetrieval` function. Default NULL.}

\item{tissueSpecificGeneType}{An integer describing type of tissue specific genes to be used, 1 for "All", 2 for "Tissue-Enriched",3 for "Tissue-Enhanced", and 4 for "Group-Enriched". Default 1.}

\item{multiHypoCorrection}{Flag to carry out multiple hypothesis correction. Default TRUE.}
}
\value{
A list object with three objects, first is the enrichment matrix, second is the list containing the tissue-specific genes found in the input genes, third is the vector containing not found genes.
}
\description{
Calculation of tissue specific gene enrichment using hypergeometric test for custom datasets
}
\author{
Ashish Jain
}
