% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissueSpecificGenes}
\alias{tissueSpecificGenes}
\title{Tissue specific gene}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Gene}{Gene Name, character}
  \item{Group}{Tissue specific group, character}
  \item{Tissue}{Tissue Name, character}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
tissueSpecificGenes
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
