% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueEnrich.R
\name{tissueSpecificGenesRetrieval}
\alias{tissueSpecificGenesRetrieval}
\title{Calculation of tissue specific genes by using the algorithm from human protein atlas project.}
\usage{
tissueSpecificGenesRetrieval(expressionData, foldChangeThreshold = 5,
  maxNumberOfTissues = 7, expressedGeneThreshold = 1)
}
\arguments{
\item{expressionData}{A dataframe object containing gene expression values (Rows are genes and Tissues are columns) .}

\item{foldChangeThreshold}{A number. Threshold of fold change, default 5.}

\item{maxNumberOfTissues}{A number. Maximum number of tissues in a group for group enriched genes, default 7.}

\item{expressedGeneThreshold}{A number. Minimum gene expression cutoff for the gene to be called as expressed, default 1.}
}
\value{
A data frame object with three columns, Gene, Tissues, and the enrichment group of the gene in the given tissue.
}
\description{
Calculation of tissue specific genes by using the algorithm from human protein atlas project.
}
\examples{
library(tidyverse)
data<-read.table("humanProteinAtlasV18.tsv",header = T,sep = "\\t")
spreadData<-data \%>\% spread(key = Sample,value= Value)
row.names(spreadData)<-spreadData[,1]
humanProteinAtlas<-spreadData[gene,c(4:ncol(spreadData))]
TSgenes<-tissueSpecificGenesRetrieval(humanProteinAtlas)
head(TSgenes)
}
