% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueEnrich.R
\name{teEnrichmentCustom}
\alias{teEnrichmentCustom}
\title{Calculation of tissue-specific gene enrichment using hypergeometric test for custom datasets}
\usage{
teEnrichmentCustom(inputGenes = NULL, tissueSpecificGenes = NULL,
  tissueSpecificGeneType = 1, multiHypoCorrection = TRUE)
}
\arguments{
\item{inputGenes}{A vector containing the input genes.}

\item{tissueSpecificGenes}{A dataframe object. Output from `teGeneRetrieval` function. Default NULL.}

\item{tissueSpecificGeneType}{An integer describing the type of tissue-specific genes to be used. 1 for "All" (default), 2 for "Tissue-Enriched",3 for "Tissue-Enhanced", and 4 for "Group-Enriched". Default 1.}

\item{multiHypoCorrection}{Flag to correct P-values for multiple hypothesis using BH method. Default TRUE.}
}
\value{
A list object with three items, first is the enrichment matrix,
second is the list containing the tissue-specificity information
of the tissue-specific genes found in the input genes,
third is the vector containing genes not found in our RNA-Seq datasets.
}
\description{
This function calculates the tissue-specific gene enrichment in the input gene set using the custom tissue-specific genes dataset.
}
\examples{
library(dplyr)
library(ggplot2)
data<-system.file("extdata", "test.expressiondata.txt", package = "TissueEnrich")
expressionData<-read.table(data,header=TRUE,row.names=1,sep='\\t')
TSgenes<-teGeneRetrieval(expressionData)
head(TSgenes)
genes<-system.file("extdata", "inputGenesEnsembl.txt", package = "TissueEnrich")
inputGenes<-scan(genes,character())
output<-teEnrichmentCustom(inputGenes,TSgenes)
#Plotting the P-Values
ggplot(output[[1]],aes(x=reorder(Tissue,-Log10PValue),y=Log10PValue,label = Tissue.Specific.Genes,fill = Tissue))+
geom_bar(stat = 'identity')+
labs(x='', y = '-LOG10(P-Value)')+
theme_bw()+
theme(legend.position="none")+
theme(plot.title = element_text(hjust = 0.5,size = 20),axis.title = element_text(size=15))+
theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1),
panel.grid.major= element_blank(),panel.grid.minor = element_blank())
}
\author{
Ashish Jain, Geetu Tuteja
}
