% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissueEnrich.R
\name{tissueSpecificGeneEnrichmentCustom}
\alias{tissueSpecificGeneEnrichmentCustom}
\title{Calculation of tissue specific gene enrichment using hypergeometric test for custom datasets}
\usage{
tissueSpecificGeneEnrichmentCustom(inputGenes = NULL,
  tissueSpecificGenes = NULL, tissueSpecificGeneType = 1,
  multiHypoCorrection = TRUE)
}
\arguments{
\item{inputGenes}{A vector containing the input genes.}

\item{tissueSpecificGenes}{A dataframe object. Output from `tissueSpecificGenesRetrieval` function. Default NULL.}

\item{tissueSpecificGeneType}{An integer describing type of tissue specific genes to be used, 1 for "All", 2 for "Tissue-Enriched",3 for "Tissue-Enhanced", and 4 for "Group-Enriched". Default 1.}

\item{multiHypoCorrection}{Flag to carry out multiple hypothesis correction. Default TRUE.}
}
\value{
A list object with three objects, first is the enrichment matrix, second is the list containing the tissue-specific genes found in the input genes, third is the vector containing genes not found in our data.
}
\description{
Calculation of tissue specific gene enrichment using hypergeometric test for custom datasets
}
\examples{
library(tidyverse)
data<-system.file("extdata", "combined-proteincodingGenedataCombine.txt", package = "TissueEnrich")
expressionData<-read.table(data,header=TRUE,row.names=1,sep='\\t')
TSgenes<-tissueSpecificGenesRetrieval(expressionData)
head(TSgenes)
genes<-system.file("extdata", "inputGenesEnsembl.txt", package = "TissueEnrich")
inputGenes<-scan(genes,character())
output<-tissueSpecificGeneEnrichmentCustom(inputGenes,TSgenes)
library(plotly)
plot_ly(output[[1]], x = ~reorder(Tissue,-Log10PValue), y = ~Log10PValue, type = "bar",
  source = "select", color = ~Tissue,
  hoverinfo = 'text',height = 700, text = ~paste('</br> Tissue Name: ',Tissue,
  '</br> -Log10(P-Value): ', Log10PValue,
  '</br> Tissue Specific Genes: ', Tissue.Specific.Genes))
}
\author{
Ashish Jain, Geetu Tuteja
}
